+++ src/views/canvas/CanvasView.vue
@@ -1,6 +1,6 @@
 <template>
   <v-app>
-    
+    <ChannelCommonMenu />
     <h1>canvas View Test</h1>
     <div>
       <CanvasListComponent />
@@ -10,11 +10,13 @@
 </template>
 
 <script>
+import ChannelCommonMenu from "@/components/basic/ChannelCommonMenu.vue";
 import CanvasListComponent from "@/components/canvas/CanvasListComponent.vue";
 import CanvasDetailComponent from "@/components/canvas/CanvasDetailComponent.vue";
 
 export default {
   components: {
+    ChannelCommonMenu,
     CanvasListComponent,
     CanvasDetailComponent,
   },
